+++ internal/merchantProcessor/localMerchantProcessor.go
@@ -2,6 +2,7 @@ package merchantProcessor
 
 import (
 	"errors"
+
 	"github.com/KerbsOD/TusLibros/internal/creditCard"
 )
 
@@ -13,7 +14,7 @@ func NewLocalMerchantProcessor() *LocalMerchantProcessor {
 	return &LocalMerchantProcessor{"1111222233334444"}
 }
 
-func (lmp *LocalMerchantProcessor) DebitOn(anAmount float64, aCreditCard *creditCard.CreditCard) error {
+func (lmp *LocalMerchantProcessor) DebitOn(anAmount int, aCreditCard *creditCard.CreditCard) error {
 	if aCreditCard.Number() != lmp.validCreditCardNumber {
 		return errors.New(InvalidCreditCardErrorMessage)
 	}
