+++ internal/merchantProcessor/merchantProcessor.go
@@ -3,5 +3,5 @@ package merchantProcessor
 import "github.com/KerbsOD/TusLibros/internal/creditCard"
 
 type MerchantProcessor interface {
-	DebitOn(anAmount float64, aCreditCard *creditCard.CreditCard) error
+	DebitOn(anAmount int, aCreditCard *creditCard.CreditCard) error
 }
